requirements.txt
websockets>=11.0.3
pyyaml>=6.0
requests>=2.31.0
asyncio